+++ src/Components/Routes/PrivateRoute.js
@@ -2,11 +2,11 @@ import React from 'react'
 import { Route, Redirect } from 'react-router-dom'
 import { useAuth } from '../Context/AuthContext'
 
-function PrivateRoute({component : Component, changeTheme, ...rest }) {
+function PrivateRoute({component : Component, ...rest }) {
     const { currentUser } = useAuth()
     return (
         <Route {...rest} 
-        render={ props => currentUser ? <Component {...props} changeTheme={changeTheme}  /> : <Redirect to="/" />} >
+        render={ props => currentUser ? <Component {...props}  /> : <Redirect to="/" />} >
 
         </Route>
     )
